+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,5 @@
 *   More nvm methods
+
+<!---->
+
+*   Requirements IPipelineHookContext